Misc fixes and improvements
It is an anti-pattern to pass an async function to the `Promise` constructor, because errors thrown inside the async function do not fail the wrapper Promise. This PR changes a manually constructed promise inside `_runTopology` to be an async function instead.

When running a topology, there may be nodes that fail while other nodes are still running. Right now, we fail fast. This PR changes that so we wait for all the promises to resolve, which is more consistent with how an event loop behaves.

Now since many nodes can be in an errored state, we removed the top-level `error` property on the snapshot and instead set an `error` property per-node.